Add tooling header
