+++ src/tests.cmake
@@ -52,6 +52,7 @@ target_sources(testprecice
     src/mesh/tests/VertexTest.cpp
     src/partition/tests/ProvidedPartitionTest.cpp
     src/partition/tests/ReceivedPartitionTest.cpp
+    src/partition/tests/fixtures.hpp
     src/precice/tests/DataContextTest.cpp
     src/precice/tests/ParallelTests.cpp
     src/precice/tests/SerialTests.cpp
@@ -68,9 +69,9 @@ target_sources(testprecice
     src/testing/TestContext.hpp
     src/testing/Testing.cpp
     src/testing/Testing.hpp
-    src/testing/main.cpp
     src/testing/WaveformFixture.cpp
     src/testing/WaveformFixture.hpp
+    src/testing/main.cpp
     src/testing/tests/ExampleTests.cpp
     src/time/tests/WaveformTest.cpp
     src/utils/tests/AlgorithmTest.cpp
